+++ R/khsmisc.R
@@ -47,6 +47,8 @@
 #'
 #' \code{\link[khsmisc]{table1}}: Formatted stratified descriptive tables using gtsummary
 #'
+#' \code{\link[khsmisc]{table2}}: Stratified result tables
+#'
 #' \code{\link[khsmisc]{rates}}: Event counts, person-time, incidence rates, and cumulative incidence
 #'
 #' @section Functions for plots:
@@ -54,6 +56,8 @@
 #' \code{\link[khsmisc]{stripplot}}: Box-whiskers/dot plot
 #'
 #' \code{\link[khsmisc]{corrmat}}: Correlation matrix
+#'
+#' \code{\link[khsmisc]{exclusion_flowchart}}: Flowcharts of exclusions from study
 #'
 #' @docType package
 #' @name khsmisc