+++ R/plots.R
@@ -268,3 +268,100 @@ corrmat <- function(
     ggplot2::guides(fill = ggplot2::guide_colorbar(barwidth = 1, barheight = 5,
                                                    title.position = "top", title.hjust = 0.5))
 }
+
+
+#' Exclusion Flowchart using DiagrammeR
+#'
+#' @description
+#' This function plots a flow chart using \code{\link[DiagrammeR]{grViz}}.
+#' It is restricted to the simple case of sequential exclusions from a single
+#' study population and not suited for CONSORT flowcharts for a
+#' parallel-group study (i.e., a randomized-controlled trial).
+#' The \code{left} column describes the flow of the group that ends up
+#' being included. The \code{right} column describes exclusions for various
+#' reasons.
+#'
+#' @param design Tibble with the following columns:
+#'   * \code{left} Text for the left box of each row.
+#'   * \code{n_left} Count to be shown as 'n =' in the bottom
+#'     of the left box. To skip, use \code{NA_integer_}.
+#'   * \code{right} Text for the right box of each row.
+#'   * \code{n_right} Count to be shown as 'n =' in the bottom
+#'     of the right box. To skip, use \code{NA_integer_}.
+#' @param width Minimum width for all boxes.
+#'
+#' @details
+#' Note \code{\link[DiagrammeR]{grViz}} does not automatically
+#' generate line breaks. To avoid extra-wide boxes, manually
+#' supply line breaks using \code{\\n}. See example.
+#'
+#' @return Plot rendered using \code{\link[DiagrammeR]{grViz}}.
+#'   For a \code{design} with three rows,
+#'   the following pseudo-code would be generated: \preformatted{
+#'   grViz(paste0("digraph flowchart {
+#'     node [fontname = Helvetica, shape = rectangle, width = 4]
+#'     tab1a [label = 'left[1] \\nn = ", n_left[1],  "']
+#'     tab1b [label = 'right[1]\\nn = ", n_right[1], "']
+#'     tab2a [label = 'left[2] \\nn = ", n_left[2],  "']
+#'     tab2b [label = 'right[2]\\nn = ", n_right[2], "']
+#'     tab3a [label = 'left[3] \\nn = ", n_left[3],  "']
+#'
+#'     { rank = same; tab1a; tab1b; }
+#'     { rank = same; tab2a; tab2b; }
+#'
+#'     tab1a -> tab2a -> tab3a;
+#'     tab1a -> tab1b;
+#'     tab2a -> tab2b;
+#'     }"))
+#'   }
+#' @export
+#'
+#' @examples
+#' # Generate a flow chart for two steps of exclusions:
+#' design <- tribble(
+#'   ~left,               ~n_left, ~right,              ~n_right,
+#'   "Study base",        1000,    "Not sampled",       250,
+#'   "Study population",  750,     "Participants with\nmissing exposure data", 100,
+#'   "Complete-case set", 650,     "",                 NA_integer_)
+#' exclusion_flowchart(design, width = 2)
+
+exclusion_flowchart <- function(design, width = 4) {
+  labels <- design %>%
+    dplyr::mutate(
+      index = dplyr::row_number(),
+      text = purrr::pmap_chr(
+        .l = list(.data$index, .data$left, .data$n_left),
+        .f = ~paste0("tab", ..1, "a [label = '", ..2,
+                     dplyr::if_else(!is.na(..3),
+                                    true = paste0("\\nn = ", ..3),
+                                    false = ""),
+                     "']")),
+      text = purrr::pmap_chr(
+        .l = list(.data$index, .data$text, .data$right, .data$n_right),
+        .f = ~dplyr::if_else(..3 == "",
+                             true = ..2,
+                             false = paste0(..2, "\ntab", .x, "b [label = '", ..3,
+                                            dplyr::if_else(!is.na(..4),
+                                                           true = paste0("\\nn = ", ..4),
+                                                           false = ""),
+                                            "']"))))
+  labels <- paste(labels$text, sep = "\n", collapse = "\n")
+
+  # Rows that have a right box
+  has_right <- design %>%
+    mutate(row_number = row_number()) %>%
+    filter(.data$right != "") %>%
+    pull(.data$row_number)
+
+  DiagrammeR::grViz(paste0(
+    "digraph flowchart {
+      node [fontname = Helvetica, shape = rectangle, width = ", width, "]\n",
+    labels,
+    paste0("{ rank = same; tab", has_right, "a; tab", has_right, "b; }",
+           sep = "\n", collapse = "\n"),
+    paste0(paste0("tab", 1:(nrow(design)-1), "a -> ", collapse = " "),
+           "tab", nrow(design), "a;\n"),
+    paste0("tab", has_right, "a -> tab", has_right, "b;",
+           sep = "\n", collapse = "\n"),
+    "}"))
+}