+++ R/tables.R
@@ -326,12 +326,16 @@ mygt <- function(df) {
 #' data(mtcars)
 #'
 #' # Stratify by "gear":
+#' \dontrun{
 #' mtcars %>%
 #'   table1(by = gear)
+#' }
 #'
 #' # Show only selected variables in rows:
+#' \dontrun{
 #' mtcars %>%
 #'   table1(cyl, mpg, disp, by = gear)
+#' }
 table1 <- function(data,
                    ...,
                    by        = NULL,